requirements.txt
discord.py == 1.3.3
python-twitter == 3.5
requests == 2.23.0